Implemented held notes
Implemented held note hitting, scoring, held notes change color if released mid-hold

+++ Assets/Scripts/ScoreManager.cs
@@ -4,10 +4,10 @@ using UnityEngine.UI;
 public class ScoreManager : MonoBehaviour
 {
     //public Text scoreText;  // The UI Text object that will display the score.
-    private int score = 0;  // The player's score.
+    private float score = 0;  // The player's score.
 
     // This function will be called to update the score.
-    public void AddScore(int points)
+    public void AddScore(float points)
     {
         score += points * ComboCounter.Instance.GetCurrentMultiplier();
         UpdateScoreDisplay();
